requirements.txt
# requirements_gemini.txt - Pure Gemini RAG System

# Core web framework
streamlit>=1.28.0
flask
flask-cors

# Google AI
google-generativeai

# Vector database
chromadb>=0.4.0

# Data processing
numpy>=1.24.0
pandas

# Google Sheets integration
gspread>=5.10.0
oauth2client>=4.1.3

# Web Search
google-search-results>=2.4.2

# Utilities
python-dotenv
requests>=2.31.0

# Optional: For better JSON handling
ujson>=5.8.0

# Optional: For async operations (if needed)
asyncio-throttle>=1.0.2

# Development
pytest>=7.4.0  # For testing
black>=23.0.0  # Code formatting